COMMON: Add getHumanReadableBytes() in util.h
This function was used in cloud-related DownloadDialog before,
and now it is also used in Options > Cloud tab.
