+++ gui/downloaddialog.cpp
@@ -24,6 +24,7 @@
 #include "backends/cloud/cloudmanager.h"
 #include "common/config-manager.h"
 #include "common/translation.h"
+#include "common/util.h"
 #include "engines/metaengine.h"
 #include "gui/browser.h"
 #include "gui/chooser.h"
@@ -205,43 +206,6 @@ void DownloadDialog::handleTickle() {
 void DownloadDialog::reflowLayout() {
 	Dialog::reflowLayout();
 	refreshWidgets();
-}
-
-namespace {
-Common::String getHumanReadableBytes(uint64 bytes, Common::String &unitsOut) {
-	Common::String result = Common::String::format("%lu", bytes);
-	unitsOut = "B";
-
-	if (bytes >= 1024) {
-		bytes /= 1024;
-		result = Common::String::format("%lu", bytes);
-		unitsOut = "KB";
-	}
-
-	double floating = bytes;
-
-	if (bytes >= 1024) {
-		bytes /= 1024;
-		floating /= 1024.0;
-		unitsOut = "MB";
-	}
-
-	if (bytes >= 1024) {
-		bytes /= 1024;
-		floating /= 1024.0;
-		unitsOut = "GB";
-	}
-
-	if (bytes >= 1024) { // woah
-		bytes /= 1024;
-		floating /= 1024.0;
-		unitsOut = "TB";
-	}
-
-	// print one digit after floating point
-	result = Common::String::format("%.1f", floating);
-	return result;
-}
 }
 
 Common::String DownloadDialog::getSizeLabelText() {
